+++ DSCResources/MSFT_CertificateImport/MSFT_CertificateImport.psm1
@@ -252,15 +252,9 @@ function Set-TargetResource
                 Verbose           = $VerbosePreference
             }
 
-            # If the built in PKI cmdlet exists then use that, otherwise command in Common module.
-            if (Test-CommandExists -Name 'Import-Certificate')
-            {
-                Import-Certificate @importCertificateParameters
-            }
-            else
-            {
-                Import-CertificateEx @importCertificateParameters
-            }
+            # Using Import-CertificateEx instead of Import-Certificate due to the following issue:
+            # https://github.com/PowerShell/CertificateDsc/issues/161
+            Import-CertificateEx @importCertificateParameters
         }
     }
     elseif ($Ensure -ieq 'Absent')